Harmonised namespaces and moved files
