+++ CMakeLists.txt
@@ -13,7 +13,12 @@ file(GLOB OBJ_SRC_FILES src/dts_renderable_shape.cpp src/dts-to-obj/*.cpp)
 file(GLOB JSON_SRC_FILES src/json-to-dts/*.cpp)
 file(GLOB VOL_SRC_FILES src/resource/*.cpp src/unvol/*.cpp)
 file(GLOB MIS_SRC_FILES src/mis-to-json/*.cpp)
-file(GLOB DTS_VIEWER_SRC_FILES src/resource/*.cpp src/dts_renderable_shape.cpp src/3space-studio/utility.cpp src/3space-studio/views/*.cpp src/3space-studio/*.cpp)
+file(GLOB DTS_VIEWER_SRC_FILES
+        src/*.cpp
+        src/content/*.cpp
+        src/content/**/*.cpp
+        src/resource/*.cpp
+        src/3space-studio/views/*.cpp src/3space-studio/*.cpp)
 
 #add_executable(mdl-to-json ${MDL_SRC_FILES})
 #add_executable(dts-to-json ${DTS_SRC_FILES})
