+++ src/content/dts/darkstar.cpp
@@ -1,24 +1,18 @@
-#ifndef DARKSTARDTSCONVERTER_DTS_IO_HPP
-#define DARKSTARDTSCONVERTER_DTS_IO_HPP
+#include "darkstar.hpp"
+#include "content/binary_io.hpp"
 
-#include <sstream>
-#include "dts_structures.hpp"
-#include "binary_io.hpp"
-
-namespace darkstar::dts
+namespace studio::content::dts::darkstar
 {
-  inline bool is_darkstar_dts(std::basic_istream<std::byte>& stream)
+  bool is_darkstar_dts(std::basic_istream<std::byte>& stream)
   {
-    using namespace binary::io;
-
     auto starting_point = stream.tellg();
 
-    dts::tag_header file_header = {
-      read<sizeof(dts::file_tag)>(stream),
-      read<dts::file_info>(stream)
+    tag_header file_header = {
+      read<sizeof(file_tag)>(stream),
+      read<file_info>(stream)
     };
 
-    if (file_header.tag != dts::pers_tag)
+    if (file_header.tag != pers_tag)
     {
       stream.seekg(starting_point, std::ios::beg);
       return false;
@@ -28,29 +22,27 @@ namespace darkstar::dts
 
     stream.seekg(starting_point, std::ios::beg);
 
-    return file_header.class_name == dts::shape::v2::shape::type_name;
+    return file_header.class_name == shape::v2::shape::type_name;
   }
 
-  inline dts::tag_header read_object_header(std::basic_istream<std::byte>& stream)
+  tag_header read_object_header(std::basic_istream<std::byte>& stream)
   {
-    using namespace binary::io;
-
-    dts::tag_header file_header = {
-      read<sizeof(dts::file_tag)>(stream),
-      read<dts::file_info>(stream)
+    tag_header file_header = {
+      read<sizeof(file_tag)>(stream),
+      read<file_info>(stream)
     };
 
-    if (file_header.tag != dts::pers_tag)
+    if (file_header.tag != pers_tag)
     {
       std::stringstream msg;
       msg << "There was an error trying to parse a portion of the DTS/DML file at byte number " << stream.tellg() << ". ";
-      msg << "Expected the " << std::string(reinterpret_cast<const char*>(dts::pers_tag.data()), dts::pers_tag.size()) << " file header to be present but it was not found.\n";
+      msg << "Expected the " << std::string(reinterpret_cast<const char*>(pers_tag.data()), pers_tag.size()) << " file header to be present but it was not found.\n";
 
       throw std::invalid_argument(msg.str());
     }
 
     file_header.class_name = read_string(stream, file_header.file_info.class_name_length);
-    file_header.version = read<dts::version>(stream);
+    file_header.version = read<version>(stream);
 
     return file_header;
   }
@@ -58,8 +50,7 @@ namespace darkstar::dts
   template<typename ShapeType>
   void read_meshes(ShapeType& shape, std::size_t num_meshes, std::basic_istream<std::byte>& stream)
   {
-    using namespace binary::io;
-    using namespace dts::mesh;
+    using namespace mesh;
     shape.meshes.reserve(num_meshes);
 
     for (auto i = 0u; i < num_meshes; ++i)
@@ -123,10 +114,9 @@ namespace darkstar::dts
     }
   }
 
-  inline dts::material_list_variant read_material_list(const dts::tag_header& object_header, std::basic_istream<std::byte>& stream)
+  material_list_variant read_material_list(const tag_header& object_header, std::basic_istream<std::byte>& stream)
   {
-    using namespace binary::io;
-    using namespace dts::material_list;
+    using namespace material_list;
     if (object_header.class_name != v2::material_list::type_name)
     {
       throw std::invalid_argument("The object was not a material list as expected.");
@@ -168,7 +158,7 @@ namespace darkstar::dts
   template<typename ShapeType>
   void read_materials(ShapeType& shape, std::basic_istream<std::byte>& stream)
   {
-    if (auto has_material_list = binary::io::read<dts::shape::v2::has_material_list_flag>(stream); has_material_list == 1)
+    if (auto has_material_list = read<shape::v2::has_material_list_flag>(stream); has_material_list == 1)
     {
       auto object_header = read_object_header(stream);
 
@@ -179,7 +169,6 @@ namespace darkstar::dts
   template<typename ShapeType>
   ShapeType read_shape_impl(std::basic_istream<std::byte>& stream)
   {
-    using namespace binary::io;
     auto header = read<decltype(ShapeType::header)>(stream);
     ShapeType shape{
       header,
@@ -209,9 +198,9 @@ namespace darkstar::dts
   }
 
 
-  inline dts::shape_variant read_shape(std::basic_istream<std::byte>& stream, std::optional<dts::tag_header> file_header)
+  shape_variant read_shape(std::basic_istream<std::byte>& stream, std::optional<tag_header> file_header)
   {
-    using namespace dts::shape;
+    using namespace shape;
     file_header = !file_header.has_value() ? read_object_header(stream) : file_header;
 
     if (file_header->class_name != v2::shape::type_name)
@@ -257,12 +246,12 @@ namespace darkstar::dts
   }
 
 
-  inline dts::shape_or_material_list read_shape(std::basic_istream<std::byte>& stream)
+  shape_or_material_list read_shape(std::basic_istream<std::byte>& stream)
   {
-    using namespace dts::shape;
-    dts::tag_header file_header = read_object_header(stream);
+    using namespace shape;
+    tag_header file_header = read_object_header(stream);
 
-    if (file_header.class_name == dts::material_list::v2::material_list::type_name)
+    if (file_header.class_name == material_list::v2::material_list::type_name)
     {
       return read_material_list(file_header, stream);
     }
@@ -273,10 +262,9 @@ namespace darkstar::dts
   template<typename RootType>
   void write_header(std::basic_ostream<std::byte>& stream, const RootType& root)
   {
-    using namespace binary::io;
     constexpr static auto empty = std::byte{ '\0' };
     boost::endian::little_uint32_t size_in_bytes{};
-    write(stream, dts::pers_tag);
+    write(stream, pers_tag);
     // This is a placeholder for the real size which will come later
     write(stream, size_in_bytes);
 
@@ -295,21 +283,19 @@ namespace darkstar::dts
     write(stream, version);
   }
 
-  inline void write_size(std::basic_ostream<std::byte>& stream, std::optional<std::uint32_t> start_offset = std::nullopt)
+  void write_size(std::basic_ostream<std::byte>& stream, std::optional<std::uint32_t> start_offset = std::nullopt)
   {
     boost::endian::little_uint32_t size_in_bytes{};
 
-    start_offset = start_offset.has_value() ? start_offset.value() + static_cast<std::uint32_t>(dts::pers_tag.size()) : static_cast<std::uint32_t>(dts::pers_tag.size());
+    start_offset = start_offset.has_value() ? start_offset.value() + static_cast<std::uint32_t>(pers_tag.size()) : static_cast<std::uint32_t>(pers_tag.size());
     std::uint32_t end_offset = static_cast<std::uint32_t>(stream.tellp());
 
     // sort out the size we reserved earlier
     size_in_bytes = end_offset - start_offset.value() - sizeof(size_in_bytes);
     stream.seekp(start_offset.value(), std::ios_base::beg);
 
-    binary::io::write(stream, size_in_bytes);
+    write(stream, size_in_bytes);
 
     stream.seekp(end_offset, std::ios_base::beg);
   }
-}// namespace darkstar::dts
-
-#endif//DARKSTARDTSCONVERTER_DTS_IO_HPP
+}
