+++ src/content/dts/darkstar_structures.hpp
@@ -5,36 +5,16 @@
 #include <array>
 #include <string>
 #include <vector>
+#include "content/3d_structures.hpp"
 #include "endian_arithmetic.hpp"
+#include "shared.hpp"
 
-namespace darkstar::dts
+namespace studio::content::dts::darkstar
 {
   namespace endian = boost::endian;
   using file_tag = std::array<std::byte, 4>;
 
-  template<std::size_t Size>
-  constexpr std::array<std::string_view, Size> make_keys(const char*(&&keys)[Size])
-  {
-    std::array<std::string_view, Size> result;
-    for (auto i = 0; i < Size; i++)
-    {
-      result[i] = keys[i];
-    }
-    return result;
-  }
-
-  constexpr file_tag to_tag(const std::array<std::uint8_t, 4> values)
-  {
-    file_tag result{};
-
-    for (auto i = 0u; i < values.size(); i++)
-    {
-      result[i] = std::byte{ values[i] };
-    }
-    return result;
-  }
-
-  constexpr file_tag pers_tag = to_tag({ 'P', 'E', 'R', 'S' });
+  constexpr file_tag pers_tag = shared::to_tag<4>({ 'P', 'E', 'R', 'S' });
 
   using version = endian::little_uint32_t;
 
@@ -46,103 +26,11 @@ namespace darkstar::dts
 
   struct tag_header
   {
-    dts::file_tag tag;
-    dts::file_info file_info;
+    file_tag tag;
+    file_info file_info;
     std::string class_name;
-    dts::version version;
+    version version;
   };
-
-  struct vector3f
-  {
-    constexpr static auto keys = make_keys({ "x", "y", "z" });
-    float x;
-    float y;
-    float z;
-  };
-
-  inline vector3f operator+(const vector3f& left, const vector3f& right)
-  {
-    return { left.x + right.x, left.y + right.y, left.z + right.z };
-  }
-
-  inline vector3f operator-(const vector3f& left, const vector3f& right)
-  {
-    return { left.x - right.x, left.y - right.y, left.z - right.z };
-  }
-
-  inline vector3f operator+=(vector3f& left, const vector3f& right)
-  {
-    left = left + right;
-    return left;
-  }
-
-  inline vector3f operator*(const vector3f& left, const vector3f& right)
-  {
-    return { left.x * right.x, left.y * right.y, left.z * right.z };
-  }
-
-  inline vector3f operator*=(vector3f& left, const vector3f& right)
-  {
-    left = left * right;
-    return left;
-  }
-
-  struct vector3f_pair
-  {
-    constexpr static auto keys = make_keys({ "min", "max" });
-    vector3f min;
-    vector3f max;
-  };
-
-  static_assert(sizeof(vector3f) == sizeof(std::array<float, 3>));
-
-  struct quaternion4s
-  {
-    constexpr static auto keys = make_keys({ "x", "y", "z", "w" });
-    endian::little_int16_t x;
-    endian::little_int16_t y;
-    endian::little_int16_t z;
-    endian::little_int16_t w;
-  };
-
-  struct quaternion4f
-  {
-    constexpr static auto keys = make_keys({ "x", "y", "z", "w" });
-    float x;
-    float y;
-    float z;
-    float w;
-  };
-
-  inline quaternion4f to_float(const quaternion4f& other)
-  {
-    return other;
-  }
-
-  inline quaternion4f to_float(const quaternion4s& other)
-  {
-    constexpr std::int16_t max = SHRT_MAX;
-
-    return {
-      float(other.x) / max,
-      float(other.y) / max,
-      float(other.z) / max,
-      float(other.w) / max
-    };
-  }
-
-  static_assert(sizeof(quaternion4s) == sizeof(std::array<endian::little_int16_t, 4>));
-
-  struct rgb_data
-  {
-    constexpr static auto keys = make_keys({ "red", "green", "blue", "rgbFlags" });
-
-    std::uint8_t red;
-    std::uint8_t green;
-    std::uint8_t blue;
-    std::uint8_t rgb_flags;
-  };
-
 
   namespace mesh::v1
   {
@@ -201,12 +89,7 @@ namespace darkstar::dts
       return result;
     }
 
-    struct texture_vertex
-    {
-      constexpr static auto keys = make_keys({ "x", "y" });
-      float x;
-      float y;
-    };
+    using texture_vertex = studio::content::texture_vertex;
 
     struct face
     {
