+++ src/dts-to-json/convert_dts.cpp
@@ -3,7 +3,7 @@
 #include <algorithm>
 #include <execution>
 #include <bitset>
-#include "json_boost.hpp"
+#include "content/json_boost.hpp"
 #include "complex_serializer.hpp"
 #include "shared.hpp"
 #include "dts_io.hpp"
