+++ src/json-to-dts/convert_json.cpp
@@ -3,7 +3,7 @@
 #include <algorithm>
 #include <execution>
 #include "dts_structures.hpp"
-#include "json_boost.hpp"
+#include "content/json_boost.hpp"
 #include "complex_serializer.hpp"
 #include "shared.hpp"
 #include "dts_io.hpp"
