+++ src/mdl-to-json/convert_mdl.cpp
@@ -3,7 +3,7 @@
 #include <algorithm>
 #include <execution>
 #include <bitset>
-#include "json_boost.hpp"
+#include "content/json_boost.hpp"
 #include "shared.hpp"
 #include "binary_io.hpp"
 #include "mdl_structures.hpp"
